refactor : reply 구조를 comment domain으로 이동
