+++ src/main/java/inu/codin/codin/domain/post/domain/comment/domain/reply/dto/request/ReplyCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.post.domain.reply.dto.request;
+package inu.codin.codin.domain.post.domain.comment.domain.reply.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
